@@ -1,106 +1,111 @@
-[tool.poetry]
+[project]
+authors = [
+    {name = "Cloud Custodian Project"},
+]
+license = "Apache-2.0"
+requires-python = "<4.0.0,>=3.9.2"
+dependencies = [
+    "PyJWT<3.0.0,>=1.7.1",
+    "adal<2.0.0,>=1.2.6",
+    "applicationinsights<1.0.0,>=0.11.9",
+    "apscheduler<4.0.0,>=3.6.3",
+    "azure-cosmos<4.0.0,>=3.1.2",
+    "azure-cosmosdb-table<2.0.0,>=1.0.6",
+    "azure-functions<2.0.0,>=1.0.8",
+    "azure-graphrbac<1.0.0,>=0.61.1",
+    "azure-identity<2.0.0,>=1.5.0",
+    "azure-keyvault<5.0.0,>=4.1.0",
+    "azure-keyvault-certificates<5.0.0,>=4.2.1",
+    "azure-keyvault-keys<5.0.0,>=4.3.1",
+    "azure-keyvault-secrets<5.0.0,>=4.2.0",
+    "azure-mgmt-advisor<10.0.0,>=9.0.0",
+    "azure-mgmt-apimanagement<2.0.0,>=1.0.0",
+    "azure-mgmt-appconfiguration==0.7.0",
+    "azure-mgmt-applicationinsights<2.0.0,>=1.0.0",
+    "azure-mgmt-appplatform<9.0.0,>=8.0.0",
+    "azure-mgmt-authorization<2.0.0,>=1.0.0",
+    "azure-mgmt-automation==0.1.1",
+    "azure-mgmt-batch<16.0.0,>=15.0.0",
+    "azure-mgmt-cdn<13.0.0,>=12.0.0",
+    "azure-mgmt-cognitiveservices<12.0.0,>=11.0.0",
+    "azure-mgmt-compute<30.0.0,>=29.1.0",
+    "azure-mgmt-containerinstance<8.0.0,>=7.0.0",
+    "azure-mgmt-containerregistry<15.0.0,>=14.0.0",
+    "azure-mgmt-containerservice<16.0.0,>=15.0.0",
+    "azure-mgmt-cosmosdb<7.0.0,>=6.1.0",
+    "azure-mgmt-costmanagement<2.0.0,>=1.0.0",
+    "azure-mgmt-databricks==1.0.0b1",
+    "azure-mgmt-datafactory<2.0.0,>=1.0.0",
+    "azure-mgmt-datalake-analytics<1.0.0,>=0.5.0",
+    "azure-mgmt-datalake-store<1.0.0,>=0.5.0",
+    "azure-mgmt-desktopvirtualization<2.0.0,>=1.0.0",
+    "azure-mgmt-dns==8.0.0b1",
+    "azure-mgmt-eventgrid<11.0.0,>=10.3.0b4",
+    "azure-mgmt-eventhub<12.0.0,>=11.0.0",
+    "azure-mgmt-frontdoor<2.0.0,>=1.0.0",
+    "azure-mgmt-hdinsight<8.0.0,>=7.0.0",
+    "azure-mgmt-iothub<2.0.0,>=1.0.0",
+    "azure-mgmt-keyvault<9.0.0,>=8.0.0",
+    "azure-mgmt-kusto<3.0.0,>=2.2.0",
+    "azure-mgmt-logic<10.0.0,>=9.0.0",
+    "azure-mgmt-machinelearningservices==2.0.0b2",
+    "azure-mgmt-managementgroups==1.0.0b1",
+    "azure-mgmt-monitor<3.0.0,>=2.0.0",
+    "azure-mgmt-msi<2.0.0,>=1.0.0",
+    "azure-mgmt-network<29.0.0,>=28.0.0",
+    "azure-mgmt-policyinsights<2.0.0,>=1.0.0",
+    "azure-mgmt-rdbms==10.2.0b17",
+    "azure-mgmt-recoveryservices<3.0.0,>=2.4.0",
+    "azure-mgmt-recoveryservicesbackup<8.0.0,>=7.0.0",
+    "azure-mgmt-redhatopenshift<2.0.0,>=1.2.0",
+    "azure-mgmt-redis<15.0.0,>=14.5.0",
+    "azure-mgmt-resource<24.0.0,>=23.0.0",
+    "azure-mgmt-resourcegraph<8.0.0,>=7.0.0",
+    "azure-mgmt-search<9.0.0,>=8.0.0",
+    "azure-mgmt-security<8.0.0,>=7.0.0",
+    "azure-mgmt-servicebus<9.0.0,>=8.2.0",
+    "azure-mgmt-servicefabric<2.0.0,>=1.0.0",
+    "azure-mgmt-signalr==0.4.0",
+    "azure-mgmt-sql<2.0.0,>=1.0.0",
+    "azure-mgmt-storage<23.0.0,>=22.1.1",
+    "azure-mgmt-streamanalytics<2.0.0,>=1.0.0rc1",
+    "azure-mgmt-subscription<2.0.0,>=1.0.0",
+    "azure-mgmt-synapse<3.0.0,>=2.0.0",
+    "azure-mgmt-trafficmanager<1.0.0,>=0.51.0",
+    "azure-mgmt-web<4.0.0,>=3.0.0",
+    "azure-storage-blob<13.0.0,>=12.8.0",
+    "azure-storage-file<3.0.0,>=2.1.0",
+    "azure-storage-file-share<13.0.0,>=12.4.1",
+    "azure-storage-queue<13.0.0,>=12.1.5",
+    "c7n",
+    "click<9.0,>=8.0",
+    "cryptography>=3.4.6",
+    "distlib<1.0.0,>=0.3.0",
+    "jmespath<2.0.0,>=1.0.0",
+    "netaddr<1.0.0,>=0.7.19",
+    "requests<3.0.0,>=2.22.0",
+]
 name = "c7n_azure"
-version = "0.7.42"
+version = "0.7.44"
 description = "Cloud Custodian - Azure Support"
 readme = "readme.md"
-homepage="https://cloudcustodian.io"
-repository="https://github.com/cloud-custodian/cloud-custodian"
-documentation="https://cloudcustodian.io/docs/"
-authors = ["Cloud Custodian Project"]
-license = "Apache-2.0"
-classifiers=[
-   "License :: OSI Approved :: Apache Software License",
-   "Topic :: System :: Systems Administration",
-   "Topic :: System :: Distributed Computing"
+classifiers = [
+    "License :: OSI Approved :: Apache Software License",
+    "Topic :: System :: Systems Administration",
+    "Topic :: System :: Distributed Computing",
 ]
 
-[tool.poetry.dependencies]
-python = "^3.8"
-c7n = {path = "../..", develop = true}
+[project.urls]
+homepage = "https://cloudcustodian.io"
+repository = "https://github.com/cloud-custodian/cloud-custodian"
+documentation = "https://cloudcustodian.io/docs/"
 
-azure-mgmt-authorization = "^1.0.0"
-azure-mgmt-advisor = "^9.0.0"
-azure-mgmt-apimanagement = "^1.0.0"
-azure-mgmt-appconfiguration = "0.7.0"
-azure-mgmt-applicationinsights = "^1.0.0"
-azure-mgmt-automation = "0.1.1"
-azure-mgmt-batch = "^15.0.0"
-azure-mgmt-cognitiveservices = "^11.0.0"
-azure-mgmt-cosmosdb = "^6.1.0"
-azure-mgmt-costmanagement = "^1.0.0"
-azure-mgmt-containerinstance = "^7.0.0"
-azure-mgmt-compute = "^19.0.0"
-azure-mgmt-cdn = "^12.0.0"
-azure-mgmt-containerregistry = "^10.3.0"
-azure-mgmt-containerservice = "^15.0.0"
-azure-mgmt-databricks = "1.0.0b1"
-azure-mgmt-datalake-analytics = "^0.5.0"
-azure-mgmt-datalake-store = "^0.5.0"
-azure-mgmt-datafactory = "^1.0.0"
-azure-mgmt-desktopvirtualization = "^1.0.0"
-azure-mgmt-dns = "8.0.0b1"
-azure-mgmt-eventgrid = "^10.3.0b4"
-azure-mgmt-eventhub = "^11.0.0"
-azure-mgmt-hdinsight = "^7.0.0"
-azure-mgmt-iothub = "^1.0.0"
-azure-mgmt-keyvault = "^8.0.0"
-azure-mgmt-kusto = "^2.2.0"
-azure-mgmt-machinelearningservices = "2.0.0b2"
-azure-mgmt-managementgroups = "1.0.0b1"
-azure-mgmt-network = "^28.0.0"
-azure-mgmt-redhatopenshift = "^1.2.0"
-azure-mgmt-resourcegraph = "^7.0.0"
-azure-mgmt-resource = "^16.0.0"
-azure-mgmt-rdbms = "10.2.0b17"
-azure-mgmt-search = "^8.0.0"
-azure-mgmt-sql = "^1.0.0"
-azure-mgmt-storage = "^17.0.0"
-azure-mgmt-streamanalytics = "^1.0.0rc1"
-azure-mgmt-subscription = "^1.0.0"
-azure-mgmt-web = "^3.0.0"
-azure-mgmt-logic = "^9.0.0"
-azure-storage-blob = "^12.8.0"
-azure-storage-queue = "^12.1.5"
-azure-cosmosdb-table = "^1.0.6"
-applicationinsights = "^0.11.9"
-apscheduler = "^3.6.3"
-distlib = "^0.3.0"
-PyJWT = ">=1.7.1,<3.0.0"
-adal = "^1.2.6"
-requests = "^2.22.0"
-azure-functions = { version = "^1.0.8", python = "~3" }
-azure-graphrbac = "^0.61.1"
-netaddr = "^0.7.19"
-azure-storage-file = "^2.1.0"
-azure-mgmt-policyinsights = "^1.0.0"
-azure-mgmt-monitor = "^2.0.0"
-click = "^8.0"
-azure-cosmos = "^3.1.2"
-azure-mgmt-redis = "^14.5.0"
-azure-keyvault-secrets = "^4.2.0"
-azure-keyvault-keys = "^4.3.1"
-azure-keyvault-certificates = "^4.2.1"
-azure-identity = "^1.5.0"
-azure-keyvault = "^4.1.0"
-azure-storage-file-share = "^12.4.1"
-cryptography = ">=3.4.6"
-azure-mgmt-msi = "^1.0.0"
-jmespath = "^1.0.0"
-azure-mgmt-servicefabric = "^1.0.0"
-azure-mgmt-trafficmanager = "^0.51.0"
-azure-mgmt-frontdoor = "^1.0.0"
-azure-mgmt-security = "^1.0.0"
-azure-mgmt-servicebus = "^8.2.0"
-azure-mgmt-appplatform = "^8.0.0"
-azure-mgmt-recoveryservices = "^2.4.0"
-azure-mgmt-signalr = "0.4.0"
-azure-mgmt-synapse = "^2.0.0"
-azure-mgmt-recoveryservicesbackup = "^7.0.0"
-
-[tool.poetry.group.dev.dependencies]
-vcrpy-unittest = ">=0.1.7"
-parameterized = ">=0.7.1"
+[dependency-groups]
+dev = [
+    "parameterized>=0.7.1",
+    "vcrpy-unittest>=0.1.7",
+]
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
