@@ -4,7 +4,9 @@ from c7n.manager import resources
 from c7n import query
 from c7n.query import QueryResourceManager
 from c7n.filters import CrossAccountAccessFilter
-from c7n.utils import local_session
+from c7n.utils import local_session, type_schema
+from c7n.tags import universal_augment
+from c7n.actions import BaseAction
 
 
 @resources.register("lex-bot")
@@ -40,6 +42,71 @@ class LexV2Bot(QueryResourceManager):
     source_mapping = {"describe": query.DescribeWithResourceTags, "config": query.ConfigSource}
 
 
+class LexV2BotAliasDescribe(query.ChildDescribeSource):
+    def augment(self, resources):
+        client = local_session(self.manager.session_factory).client('lexv2-models')
+        for r in resources:
+            botalias = client.describe_bot_alias(
+                botId=r['c7n:parent-id'], botAliasId=r['botAliasId'])
+            botalias.pop('ResponseMetadata')
+            r.update(botalias)
+        return universal_augment(self.manager, resources)
+
+    def resources(self, query):
+        resources = self.query.filter(self.manager, **query)
+        return [r for r in resources if r['botAliasStatus'] == 'Available']
+
+
+@resources.register('lexv2-bot-alias')
+class LexV2BotAlias(query.ChildResourceManager):
+    class resource_type(query.TypeInfo):
+        service = 'lexv2-models'
+        parent_spec = ('lexv2-bot', 'botId', True)
+        enum_spec = ('list_bot_aliases', 'botAliasSummaries', None)
+        name = 'botAliasId'
+        id = 'botAliasId'
+        universal_taggable = object()
+        arn = 'botAliasArn'
+        arn_service = 'lex'
+        cfn_type = 'AWS::Lex::BotAlias'
+        permission_prefix = "lex"
+        permissions_augment = ("lex:DescribeBotAlias",)
+    source_mapping = {'describe-child': LexV2BotAliasDescribe, 'config': query.ConfigSource}
+
+    def get_arns(self, resources):
+        arns = []
+        for r in resources:
+            arns.append(self.generate_arn(f"bot-alias/{r['c7n:parent-id']}/{r['botAliasId']}"))
+        return arns
+
+
+@LexV2BotAlias.action_registry.register('delete')
+class DeleteLexV2BotAlias(BaseAction):
+    """
+    Deletes LexV2 bot aliases.
+
+    :example:
+
+    .. code-block:: yaml
+
+            policies:
+              - name: delete-lex-bot-alias
+                resource: lexv2-bot-alias
+                actions:
+                  - type: delete
+    """
+    schema = type_schema('delete')
+    permissions = ('lex:DeleteBotAlias',)
+
+    def process(self, resources):
+        client = local_session(self.manager.session_factory).client('lexv2-models')
+        for r in resources:
+            client.delete_bot_alias(
+                botAliasId=r['botAliasId'],
+                botId=r['c7n:parent-id']
+            )
+
+
 @LexV2Bot.filter_registry.register('cross-account')
 class LexV2BotCrossAccountAccessFilter(CrossAccountAccessFilter):
     """Filters all LexV2 bots with cross-account access
@@ -69,6 +136,40 @@ class LexV2BotCrossAccountAccessFilter(CrossAccountAccessFilter):
             client.describe_resource_policy,
             resourceArn=self.manager.generate_arn(r['botId']),
             ignore_err_codes=('ResourceNotFoundException'))
+        if result:
+            pol = result.get('policy', None)
+            r[self.policy_attribute] = pol
+        return pol
+
+
+@LexV2BotAlias.filter_registry.register('cross-account')
+class LexV2BotAliasCrossAccountAccessFilter(CrossAccountAccessFilter):
+    """Filters all LexV2 bot aliases with cross-account access
+
+    :example:
+
+    .. code-block:: yaml
+
+            policies:
+              - name: lex-bot-alias-cross-account
+                resource: lexv2-bot-alias
+                filters:
+                  - type: cross-account
+                    whitelist_from:
+                      expr: "accounts.*.accountNumber"
+                      url: accounts_url
+    """
+    permissions = ('lex:DescribeResourcePolicy',)
+    policy_attribute = 'c7n:Policy'
+
+    def get_resource_policy(self, r):
+        client = local_session(self.manager.session_factory).client('lexv2-models')
+        pol = None
+        result = self.manager.retry(
+                client.describe_resource_policy,
+                resourceArn=self.manager.generate_arn
+                (f"bot-alias/{r['c7n:parent-id']}/{r['botAliasId']}"),
+                ignore_err_codes=('ResourceNotFoundException'))
         if result:
             pol = result.get('policy', None)
             r[self.policy_attribute] = pol
