@@ -54,10 +54,15 @@ ResourceMap = {
   "aws.codebuild-credential": "c7n.resources.code.CodeBuildSourceCredentials",
   "aws.codecommit": "c7n.resources.code.CodeRepository",
   "aws.codedeploy-app": "c7n.resources.code.CodeDeployApplication",
+  "aws.codedeploy-config": "c7n.resources.code.CodeDeployConfig",
   "aws.codedeploy-deployment": "c7n.resources.code.CodeDeployDeployment",
   "aws.codedeploy-group": "c7n.resources.code.CodeDeployDeploymentGroup",
   "aws.codepipeline": "c7n.resources.code.CodeDeployPipeline",
   "aws.composite-alarm": "c7n.resources.cw.CompositeAlarm",
+  "aws.comprehend-document-classifier": "c7n.resources.comprehend.ComprehendDocumentClassifier",
+  "aws.comprehend-endpoint": "c7n.resources.comprehend.ComprehendEndpoint",
+  "aws.comprehend-entity-recognizer": "c7n.resources.comprehend.ComprehendEntityRecognizer",
+  "aws.comprehend-flywheel": "c7n.resources.comprehend.ComprehendFlywheel",
   "aws.config-recorder": "c7n.resources.config.ConfigRecorder",
   "aws.config-rule": "c7n.resources.config.ConfigRule",
   "aws.connect-campaign": "c7n.resources.connect.ConnectCampaign",
@@ -68,6 +73,8 @@ ResourceMap = {
   "aws.datasync-agent": "c7n.resources.datasync.DataSyncAgent",
   "aws.datasync-task": "c7n.resources.datasync.DataSyncTask",
   "aws.dax": "c7n.resources.dynamodb.DynamoDbAccelerator",
+  "aws.delivery-destination": "c7n.resources.cw.DeliveryDestination",
+  "aws.destination": "c7n.resources.cw.Destination",
   "aws.devicefarm-project": "c7n.resources.devicefarm.DevicefarmProject",
   "aws.directconnect": "c7n.resources.directconnect.DirectConnect",
   "aws.directory": "c7n.resources.directory.Directory",
@@ -119,6 +126,7 @@ ResourceMap = {
   "aws.flow-log": "c7n.resources.vpc.FlowLog",
   "aws.fsx": "c7n.resources.fsx.FSx",
   "aws.fsx-backup": "c7n.resources.fsx.FSxBackup",
+  "aws.fsx-volume": "c7n.resources.fsx.FSxVolume",
   "aws.gamelift-build": "c7n.resources.gamelift.GameLiftBuild",
   "aws.gamelift-fleet": "c7n.resources.gamelift.GameLiftFleet",
   "aws.glacier": "c7n.resources.glacier.Glacier",
@@ -160,6 +168,8 @@ ResourceMap = {
   "aws.kafka-config": "c7n.resources.kafka.KafkaClusterConfiguration",
   "aws.kendra": "c7n.resources.kendra.KendraIndex",
   "aws.key-pair": "c7n.resources.vpc.KeyPair",
+  "aws.keyspace": "c7n.resources.keyspaces.Keyspace",
+  "aws.keyspace-table": "c7n.resources.keyspaces.Table",
   "aws.kinesis": "c7n.resources.kinesis.KinesisStream",
   "aws.kinesis-analytics": "c7n.resources.kinesis.AnalyticsApp",
   "aws.kinesis-analyticsv2": "c7n.resources.kinesis.KinesisAnalyticsAppV2",
@@ -172,10 +182,10 @@ ResourceMap = {
   "aws.launch-template-version": "c7n.resources.ec2.LaunchTemplate",
   "aws.lex-bot": "c7n.resources.lex.LexBot",
   "aws.lexv2-bot": "c7n.resources.lex.LexV2Bot",
+  "aws.lexv2-bot-alias": "c7n.resources.lex.LexV2BotAlias",
   "aws.lightsail-db": "c7n.resources.lightsail.Database",
   "aws.lightsail-elb": "c7n.resources.lightsail.LoadBalancer",
   "aws.lightsail-instance": "c7n.resources.lightsail.Instance",
-  "aws.log-destination": "c7n.resources.cw.LogDestination",
   "aws.log-group": "c7n.resources.cw.LogGroup",
   "aws.log-metric": "c7n.resources.cw.LogMetric",
   "aws.memorydb": "c7n.resources.memorydb.MemoryDb",
@@ -215,10 +225,13 @@ ResourceMap = {
   "aws.quicksight-group": "c7n.resources.quicksight.QuicksightGroup",
   "aws.quicksight-user": "c7n.resources.quicksight.QuicksightUser",
   "aws.r53domain": "c7n.resources.route53.Route53Domain",
+  "aws.resource-share-self": "c7n.resources.ram.ResourceShareSelf",
+  "aws.resource-share-other": "c7n.resources.ram.ResourceShareOther",
   "aws.rds": "c7n.resources.rds.RDS",
   "aws.rds-cluster": "c7n.resources.rdscluster.RDSCluster",
   "aws.rds-cluster-param-group": "c7n.resources.rdsparamgroup.RDSClusterParamGroup",
   "aws.rds-cluster-snapshot": "c7n.resources.rdscluster.RDSClusterSnapshot",
+  "aws.rds-db-shard-group": "c7n.resources.rdscluster.RDSDbShardGroup",
   "aws.rds-param-group": "c7n.resources.rdsparamgroup.RDSParamGroup",
   "aws.rds-proxy": "c7n.resources.rds.RDSProxy",
   "aws.rds-reserved": "c7n.resources.rds.ReservedRDS",
@@ -305,6 +318,7 @@ ResourceMap = {
   "aws.timestream-database": "c7n.resources.timestream.TimestreamDatabase",
   "aws.timestream-table": "c7n.resources.timestream.TimestreamTable",
   "aws.timestream-influxdb": "c7n.resources.timestream.TimestreamInfluxDB",
+  "aws.timestream-influxdb-cluster": "c7n.resources.timestream.TimestreamInfluxDBCluster",
   "aws.transfer-server": "c7n.resources.transfer.TransferServer",
   "aws.transfer-user": "c7n.resources.transfer.TransferUser",
   "aws.transit-attachment": "c7n.resources.vpc.TransitGatewayAttachment",
