@@ -3,7 +3,7 @@
 from c7n.manager import resources
 from c7n.query import ConfigSource, QueryResourceManager, TypeInfo, DescribeSource
 from c7n.tags import universal_augment
-from c7n.filters import ValueFilter
+from c7n.filters import ValueFilter, ListItemFilter
 from c7n.utils import type_schema, local_session
 
 
@@ -58,7 +58,6 @@ class DescribeWafV2(DescribeSource):
 
     def resources(self, query):
         scope = (query or {}).get('Scope', 'REGIONAL')
-
         # The AWS API does not include the scope as part of the WebACL information, but scope
         # is a required parameter for most API calls - we augment the resource with the desired
         # scope here in order to use it downstream for API calls
@@ -205,3 +204,71 @@ class WAFV2LoggingFilter(ValueFilter):
         return [
             r for r in resources if self.match(
                 r.get(self.annotation_key, {}))]
+
+
+@WAFV2.filter_registry.register('web-acl-rules')
+class WAFV2ListAllRulesFilter(ListItemFilter):
+    """
+    Return all rules inside the Web ACL, including rules in rule groups.
+    Allows filtering based on any field within the rules data.
+
+    :example:
+
+    .. code-block:: yaml
+
+        policies:
+          - name: find-rule-groups
+            resource: aws.wafv2
+            filters:
+              - type: web-acl-rules
+                attrs:
+                  - type: value
+                    key: Type
+                    value: RuleGroup
+                    op: in
+
+    """
+
+    schema = type_schema(
+        'web-acl-rules',
+        attrs={'$ref': '#/definitions/filters_common/list_item_attrs'}
+    )
+    permissions = ('wafv2:GetRuleGroup',)
+    annotate_items = True
+    item_annotation_key = 'c7n:WebACLAllRules'
+
+    def get_item_values(self, resource):
+        client = local_session(self.manager.session_factory).client(
+            'wafv2', region_name=self.manager.region
+        )
+
+        all_rules = []
+
+        for rule in resource.get('Rules', []):
+            if rule.get("Statement", {}).get('RuleGroupReferenceStatement'):
+                rule_group_arn = rule['Statement']['RuleGroupReferenceStatement']['ARN']
+                scope = resource['Scope']
+
+                rule_group_response = client.get_rule_group(
+                    Name=rule_group_arn.split('/')[-2],
+                    Id=rule_group_arn.split('/')[-1],
+                    Scope=scope
+                )
+                rule_group = rule_group_response.get('RuleGroup', {})
+
+                rule_details = {
+                    "Type": "RuleGroup",
+                    "Name": rule.get('Name'),
+                    "RuleGroupARN": rule_group_arn,
+                    "Rules": rule_group.get('Rules', [])
+                }
+                all_rules.append(rule_details)
+            else:
+                rule_details = {
+                    "Type": "Standalone",
+                    "Name": rule.get('Name'),
+                    "Rule": rule
+                }
+                all_rules.append(rule_details)
+
+        return all_rules
