@@ -493,14 +493,97 @@ class LogGroupTest(BaseTest):
         self.assertIn('c7n:MetricAlarms', resources[0])
 
 
-class LogDestinationTest(BaseTest):
-    def test_log_destination(self):
-        factory = self.replay_flight_data('test_log_destination')
+class DestinationTest(BaseTest):
+
+    def test_destination(self):
+        factory = self.replay_flight_data('test_destination')
         client = local_session(factory).client('logs')
         p = self.load_policy(
             {
-                'name': 'log-destination-tag',
-                'resource': 'log-destination',
+                'name': 'destination-tag',
+                'resource': 'destination',
+                'filters': [{
+                    'type': 'value',
+                    'key': 'destinationName',
+                    'value': 'test-destination'
+                }],
+                'actions': [{
+                    'type': 'tag',
+                    'key': 'test-tag',
+                    'value': 'test-value'
+                }]
+            },
+        session_factory=factory)
+        resources = p.run()
+        self.assertEqual(len(resources), 1)
+        self.assertEqual(resources[0]['destinationName'], 'test-destination')
+        tags = client.list_tags_for_resource(resourceArn=resources[0]['arn'])['tags']
+        assert tags['test-tag'] == 'test-value'
+
+        p = self.load_policy(
+            {
+                'name': 'destination-untag',
+                'resource': 'destination',
+                'filters': [{
+                    'type': 'value',
+                    'key': 'destinationName',
+                    'value': 'test-destination'
+                }],
+                'actions': [{
+                    'type': 'remove-tag',
+                    'tags': ['test-tag']
+                }]
+            },
+        session_factory=factory)
+        resources = p.run()
+        self.assertEqual(len(resources), 1)
+        tags = client.list_tags_for_resource(resourceArn=resources[0]['arn'])['tags']
+        assert 'test-tag' not in tags
+
+    def test_log_destination_cross_account(self):
+        factory = self.replay_flight_data('test_destination_cross_account')
+        p = self.load_policy(
+            {
+                'name': 'cross-account-destination',
+                'resource': 'destination',
+                'filters': [{
+                    'type': 'cross-account'
+                }],
+            },
+        session_factory=factory)
+        resources = p.run()
+        self.assertEqual(len(resources), 1)
+
+    def test_log_destination_delete(self):
+        factory = self.replay_flight_data('test_destination_delete')
+        client = local_session(factory).client('logs')
+        p = self.load_policy(
+            {
+                'name': 'destination-delete',
+                'resource': 'destination',
+                'filters': [{
+                    'type': 'value',
+                    'key': 'destinationName',
+                    'value': 'test-destination'
+                }],
+                'actions': ['delete']
+            },
+        session_factory=factory)
+        resources = p.run()
+        self.assertEqual(len(resources), 1)
+        destinations = client.describe_destinations(
+            DestinationNamePrefix='test-destination')['destinations']
+        self.assertEqual(len(destinations), 0)
+
+
+class DeliveryDestinationTest(BaseTest):
+    def test_log_destination(self):
+        factory = self.replay_flight_data('test_delivery_destination')
+        client = local_session(factory).client('logs')
+        p = self.load_policy(
+            {
+                'name': 'delivery-destination-tag',
+                'resource': 'delivery-destination',
                 'filters': [{
                     'type': 'value',
                     'key': 'name',
@@ -521,8 +604,8 @@ class LogDestinationTest(BaseTest):
 
         p = self.load_policy(
             {
-                'name': 'log-destination-untag',
-                'resource': 'log-destination',
+                'name': 'delivery-destination-untag',
+                'resource': 'delivery-destination',
                 'filters': [{
                     'type': 'value',
                     'key': 'name',
@@ -540,11 +623,11 @@ class LogDestinationTest(BaseTest):
         assert 'test-tag' not in tags
 
     def test_log_destination_cross_account(self):
-        factory = self.replay_flight_data('test_log_destination_cross_account')
+        factory = self.replay_flight_data('test_delivery_destination_cross_account')
         p = self.load_policy(
             {
-                'name': 'cross-account-log-destination',
-                'resource': 'log-destination',
+                'name': 'cross-account-delivery-destination',
+                'resource': 'delivery-destination',
                 'filters': [{
                     'type': 'cross-account'
                 }],
@@ -554,12 +637,12 @@ class LogDestinationTest(BaseTest):
         self.assertEqual(len(resources), 1)
 
     def test_log_destination_delete(self):
-        factory = self.replay_flight_data('test_log_destination_delete')
+        factory = self.replay_flight_data('test_delivery_destination_delete')
         client = local_session(factory).client('logs')
         p = self.load_policy(
             {
-                'name': 'log-destination-delete',
-                'resource': 'log-destination',
+                'name': 'delivery-destination-delete',
+                'resource': 'delivery-destination',
                 'filters': [{
                     'type': 'value',
                     'key': 'name',
