@@ -172,7 +172,7 @@ class TestEc2NetworkLocation(BaseTest):
         client = factory().client('ec2')
         resp = client.describe_instances()
 
-        self.assertTrue(len(resp['Reservations'][0]['Instances']), 1)
+        self.assertEqual(len(resp['Reservations'][0]['Instances']), 1)
         self.assertTrue(
             len(resp['Reservations'][0]['Instances'][0]['State']['Name']),
             'terminated'
@@ -297,6 +297,32 @@ class TestMetricFilter(BaseTest):
                 ],
             },
             session_factory=session_factory,
+        )
+        resources = policy.run()
+        self.assertEqual(len(resources), 1)
+
+    def test_metric_filter_extended_stats(self):
+        session_factory = self.replay_flight_data(
+            "test_ec2_metric_extended_stats",
+            region="us-east-2"
+        )
+        policy = self.load_policy(
+            {
+                "name": "ec2-utilization-p95",
+                "resource": "aws.ec2",
+                "filters": [
+                    {
+                        "type": "metrics",
+                        "name": "CPUUtilization",
+                        "days": 7,
+                        "value": 10,
+                        "op": "less-than",
+                        "statistics": "p95",
+                    }
+                ],
+            },
+            session_factory=session_factory,
+            config={"region": "us-east-2"},
         )
         resources = policy.run()
         self.assertEqual(len(resources), 1)
