@@ -1,51 +1,12 @@
 # Copyright The Cloud Custodian Authors.
 # SPDX-License-Identifier: Apache-2.0
 
-from azure.mgmt.security import SecurityCenter
-
-from c7n.utils import local_session
 from c7n_azure.provider import resources
 from c7n_azure.query import QueryResourceManager, QueryMeta, TypeInfo
 
 
-class DefenderResourceManager(QueryResourceManager):
-    """Manager for Microsoft Defender resources
-
-    Note: The "Microsoft Defender for Cloud" name replaces and
-    consolidates products previously called Azure Security Center
-    and Azure Defender.
-
-    The Azure Security SDK takes different arguments for its
-    SecurityCenter client than other service SDKs use.
-
-    We can override client creation here to help simplify individual
-    Defender resource definitions.
-    """
-
-    def get_client(self):
-        session = local_session(self.session_factory)
-
-        # The SecurityCenter client takes an "asc_location" parameter, and the
-        # documentation[^1] points out that this can come from the locations
-        # list (elsewhere there are references to using a subscription's
-        # "home region" for asc_location).
-        #
-        # However, from peeking at the Azure CLI's code it looks like they
-        # hardcode an arbitrary/common location[^2]. The initial pull request
-        # adding Defender to the CLI[^3] mentions that the intention is to
-        # remove asc_location from client creation and hide it from the user.
-        #
-        # Following the Azure CLI team's lead and hardcoding "centralus"
-        # here seems reasonable.
-        #
-        # [^1]: https://azuresdkdocs.blob.core.windows.net/$web/python/azure-mgmt-security/1.0.0/azure.mgmt.security.html#azure.mgmt.security.SecurityCenter  # noqa
-        # [^2]: https://github.com/Azure/azure-cli/blob/29767d75d850ddc1c24cc85bd46d861b61d77a47/src/azure-cli/azure/cli/command_modules/security/_client_factory.py#L11  # noqa
-        # [^3]: https://github.com/Azure/azure-cli/pull/7917#discussion_r238458818  # noqa
-        return SecurityCenter(session.get_credentials(), session.subscription_id, "centralus")
-
-
 @resources.register("defender-pricing")
-class DefenderPricing(DefenderResourceManager, metaclass=QueryMeta):
+class DefenderPricing(QueryResourceManager, metaclass=QueryMeta):
     """Active Microsoft Defender pricing details for supported resources.
 
     :example:
@@ -72,12 +33,18 @@ class DefenderPricing(DefenderResourceManager, metaclass=QueryMeta):
         enum_spec = ("pricings", "list", None)
         client = "SecurityCenter"
         filter_name = None
-        service = "security"
+        service = "azure.mgmt.security"
         resource_type = "Microsoft.Security/pricings"
+
+        @classmethod
+        def extra_args(cls, resource_manager):
+            return {
+                'scope_id': f'subscriptions/{resource_manager.get_session().get_subscription_id()}'
+            }
 
 
 @resources.register("defender-setting")
-class DefenderSetting(DefenderResourceManager, metaclass=QueryMeta):
+class DefenderSetting(QueryResourceManager, metaclass=QueryMeta):
     """Top-level Microsoft Defender settings for a subscription.
 
     :example:
@@ -103,12 +70,12 @@ class DefenderSetting(DefenderResourceManager, metaclass=QueryMeta):
         enum_spec = ("settings", "list", None)
         client = "SecurityCenter"
         filter_name = None
-        service = "security"
+        service = "azure.mgmt.security"
         resource_type = "Microsoft.Security/settings"
 
 
 @resources.register("defender-autoprovisioning")
-class DefenderAutoProvisioningSetting(DefenderResourceManager, metaclass=QueryMeta):
+class DefenderAutoProvisioningSetting(QueryResourceManager, metaclass=QueryMeta):
     """Auto-provisioning settings for Microsoft Defender agents.
 
     :example:
@@ -133,12 +100,12 @@ class DefenderAutoProvisioningSetting(DefenderResourceManager, metaclass=QueryMe
         enum_spec = ("auto_provisioning_settings", "list", None)
         client = "SecurityCenter"
         filter_name = None
-        service = "security"
+        service = "azure.mgmt.security"
         resource_type = "Microsoft.Security/autoProvisioningSettings"
 
 
 @resources.register("defender-alert")
-class DefenderAlertSettings(DefenderResourceManager, metaclass=QueryMeta):
+class DefenderAlertSettings(QueryResourceManager, metaclass=QueryMeta):
     """Alert settings for Microsoft Defender.
 
     :example:
@@ -163,18 +130,18 @@ class DefenderAlertSettings(DefenderResourceManager, metaclass=QueryMeta):
         enum_spec = ("security_contacts", "list", None)
         client = "securityContacts"
         filter_name = None
-        service = "security"
+        service = "azure.mgmt.security"
         resource_type = "Microsoft.Security/alertNotifications"
 
 
 @resources.register("defender-assessment")
-class DefenderAssessment(DefenderResourceManager, metaclass=QueryMeta):
+class DefenderAssessment(QueryResourceManager, metaclass=QueryMeta):
     class resource_type(TypeInfo):
         doc_groups = ["Security"]
 
         id = "id"
         name = "name"
-        service = "security"
+        service = "azure.mgmt.security"
         client = "SecurityCenter"
         enum_spec = ("assessments", "list", None)
         resource_type = 'Microsoft.Security/assessments'
@@ -188,7 +155,7 @@ class DefenderAssessment(DefenderResourceManager, metaclass=QueryMeta):
 
 
 @resources.register("defender-contact")
-class DefenderSecurityContact(DefenderResourceManager, metaclass=QueryMeta):
+class DefenderSecurityContact(QueryResourceManager, metaclass=QueryMeta):
     """Security Contacts Resource
 
     :example:
@@ -212,7 +179,7 @@ class DefenderSecurityContact(DefenderResourceManager, metaclass=QueryMeta):
 
         id = "id"
         name = "name"
-        service = "security"
+        service = "azure.mgmt.security"
         client = "SecurityCenter"
         enum_spec = ("security_contacts", "list", None)
         resource_type = "Microsoft.Security/securityContacts"
@@ -220,11 +187,11 @@ class DefenderSecurityContact(DefenderResourceManager, metaclass=QueryMeta):
 
 
 @resources.register("defender-jit-policy")
-class DefenderJitPolicy(DefenderResourceManager, metaclass=QueryMeta):
+class DefenderJitPolicy(QueryResourceManager, metaclass=QueryMeta):
     class resource_type(TypeInfo):
         doc_groups = ["Security"]
 
-        service = "security"
+        service = "azure.mgmt.security"
         client = "SecurityCenter"
         enum_spec = ("jit_network_access_policies", "list", None)
         resource_type = "Microsoft.Security/jitNetworkAccessPolicies"
