@@ -80,7 +80,7 @@ class VMInstanceViewFilter(ValueFilter):
         for host in resources:
             rg = ResourceIdParser.get_resource_group(host['id'])
             vm = ResourceIdParser.get_resource_name(host['properties']['resourceId'])
-            vmachine = client.virtual_machines.get(rg, vm)
+            vmachine = client.virtual_machines.get(rg, vm, expand='instanceView')
             if self.match(vmachine.serialize(True)):
                 matched.append(host)
         return matched
