@@ -12,6 +12,6 @@ class ArtifactRegistryRepositoryTest(BaseTest):
             session_factory=factory)
         resources = p.run()
 
-        self.assertTrue(len(resources), 1)
+        self.assertEqual(len(resources), 1)
         self.assertEqual(resources[0]['name'], 'projects/cloud-custodian/'
                                                'locations/us-central1/repositories/test')
