@@ -13,7 +13,7 @@ class PatchDeploymentTest(BaseTest):
             session_factory=factory)
         resources = p.run()
 
-        self.assertTrue(len(resources), 1)
+        self.assertEqual(len(resources), 1)
         self.assertEqual(resources[0]['name'], 'projects/cloud-custodian/patchDeployments/test')
 
         assert p.resource_manager.get_urns(resources) == [
