@@ -16,6 +16,7 @@ accept = (
     'BSD-3-Clause',
     'Apache-2.0',
     'Apache-2',
+    'PSF-2.0'
 )
 
 accept_classifiers = set(
@@ -36,24 +37,13 @@ whitelist_packages = set(
         #
         # Deps with licenses that get flagged
         'pygit2',  # tools/c7n_policystream dep, GPL w/ Linking Exception
-        'astroid',  # ci for codelint, LGPL-2.1
-        'pylint',  # ci for codelint, GPL
-        'semgrep',  # ci, LGPLv2
         'ldap3',  # mailer dependency, LGPL
         'sphinx-markdown-tables',  # docgen - GPL
         'docutils',  # docgen - couple of different licenses but bulk is public domain
-        # 'chardet',  # requests dep - LPGL
-        'websocket-client',  # c7n_kube dep - LGPL-2.1
         #
         # packages with bad metadata
-        'applicationinsights',  # MIT
-        'msal-extensions',  # MIT
         'protobuf',  # BSD-3-Clause
-        'python-http-client',  # MIT
-        'sendgrid',  # MIT
-        'typed-ast',  # apache 2.0
-        'starkbank-ecdsa',  # MIT
-        'portalocker',  # PSF
+        'uritemplate'  # Dual license BSD 3-Clause OR Apache-2.0
     )
 )
 
@@ -77,7 +67,8 @@ def main():
         delta = set(classifiers).difference(accept_classifiers)
         if (delta or not classifiers) and dname not in whitelist_packages:
             found = True
-            print(f"{dname}: {d.metadata['License']} {classifiers}")
+            license = d.metadata.get('License', None) or d.metadata.get('License-Expression', None)
+            print(f"{dname}: license:{license} classifiers:{classifiers}")
 
         seen.add(dname)
 
